% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{bwKW}
\alias{bwKW}
\title{Bandwidth selection for KW smoothing}
\usage{
bwKW(g, k = 1, minbw = 0.1)
}
\arguments{
\item{g}{KW fitted object}

\item{k}{multiplicative fudge factor}

\item{minbw}{minimum allowed value of bandwidth}
}
\description{
Bandwidth selection for KW smoothing
}
\author{
R. Koenker
}
\keyword{utility}
