% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{bwKW2}
\alias{bwKW2}
\title{Bandwidth selection for bivariate KW smoothing}
\usage{
bwKW2(g, k = 1)
}
\arguments{
\item{g}{bivariate KW fitted object}

\item{k}{multiplicative fudge factor}
}
\description{
Bandwidth selection for bivariate KW smoothing
}
\author{
R. Koenker
}
\keyword{utility}
