% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{dhuber}
\alias{dhuber}
\title{Huber density function}
\usage{
dhuber(x, mu = 0, sigma = 1, k = 1.642, heps = hubereps(k))
}
\arguments{
\item{x}{points to evaluate the density}

\item{mu}{center of symmetry of the density}

\item{sigma}{standard deviation of the nominal Gaussian model}

\item{k}{Huber k value}

\item{heps}{Huber epsilson value}
}
\value{
A vector of density values
}
\description{
Huber (1964) least favorable density for the Gaussian contamination model
}
\author{
R. Koenker
}
\keyword{utility}
