% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gosset.R
\name{Gosset}
\alias{Gosset}
\title{Gosset Criminal Finger Data}
\format{
A data frame with 3000 observations on 2 variables.
\describe{
    \item{\code{LMFinger}}{Length of Left Middle Finger (cm).}
    \item{\code{Height}}{Height (cm)}
}
}
\source{
see the man page for \code{crimtab}
}
\usage{
Gosset
}
\description{
This data was generated by dithering the cell counts
in the \code{crimtab} available in the base \pkg{stats} package.
}
\keyword{datasets}
