% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Guvenen.R
\name{Guvenen}
\alias{Guvenen}
\title{Annual Increments in Log Income}
\format{A data frame with 279 observations on two variables.
\itemize{
    \item{\code{earnings}}{annual increment in log income}
    \item{\code{logdensity}}{estimated log density values}
}}
\source{
Fatih Guvenen, Fatih Karahan, Serdar Ozkan and Jae Song, (2016)
What Do Data on Millions of U.S. Workers Reveal about Life-Cycle Earnings Dynamics?
\url{https://fguvenendotcom.files.wordpress.com/2014/04/gkos_2016_rev2_v47.pdf}
}
\usage{
Guvenen
}
\description{
Kernel density estimates of the log density of annual increments in
log income for U.S. individuals over the period 1994-2013, as estimated by \cite{Guvenen}.
}
\keyword{datasets}

