% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hubereps}
\alias{hubereps}
\title{Huber epsilon}
\usage{
hubereps(k)
}
\arguments{
\item{k}{Huber k value}
}
\value{
Huber epsilon value
}
\description{
Find the epsilon corresponding to a Huber k value
}
\author{
R. Koenker
}
