% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{KWsmooth}
\alias{KWsmooth}
\title{Smooth a Kiefer-Wolfowitz NPMLE}
\usage{
KWsmooth(f, bw = NULL, k = 2)
}
\arguments{
\item{f}{KW fitted object}

\item{bw}{bandwidth defaults to 2 * mad}

\item{k}{kernel 2 for biweight, 3 for triweight}
}
\description{
Smooth a Kiefer-Wolfowitz NPMLE
}
\author{
R. Koenker
}
\keyword{utility}
