% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Norberg.R
\name{Norberg}
\alias{Norberg}
\title{Norberg Life Insurance Data}
\format{
A data frame with 72 observations on the following 3 variables.
\itemize{
    \item{\code{OccGroup}}{Occupational Group}
    \item{\code{Exposure}}{Exposures}
    \item{\code{Deaths}}{Observed Deaths}
}
}
\usage{
Norberg
}
\description{
Norwegian Life Insurance Exposures and Claims
}
\details{
The data arise from 1125 original groups insured during all or part of
the period 1982-85 by a major Nowegian insurance company.  Exposures
can be normalized by a factor of 344 as in Hastrup (2000) and then can be
interpreted as the apriori expected number of claims (deaths) for each group.
The original 1125 groups were aggregated into 72 as in Norberg (1989).
}
\references{
Norberg, R. (1989) Experience rating in group life insurance,
Scand. Actuarial J.,194-224.

Haastrup, S. (2000) Comparison of some Bayesian analyses of heterogeneity 
in group life insurance, Scand. Actuarial J.,2-16.
}
\keyword{datasets}
