% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.GLVmix.R, R/plot.WGLVmix.R
\name{plot.GLVmix}
\alias{plot.GLVmix}
\title{Plot a GLVmix object}
\usage{
\method{plot}{GLVmix}(x, ...)

\method{plot}{GLVmix}(x, ...)
}
\arguments{
\item{x}{is the fitted object}

\item{...}{other arguments to pass to \code{symbols}, notably e.g. \code{add = TRUE}}
}
\value{
nothing (invisibly)

nothing (invisibly)
}
\description{
Given a fitted mixture model by GLVmix plot the estimated mass points

Given a fitted mixture model by GLVmix plot the estimated mass points
}
