% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medde.R
\name{plot.medde}
\alias{plot.medde}
\title{Plotting method for medde objects}
\usage{
\method{plot}{medde}(x, ...)
}
\arguments{
\item{x}{object obtained from medde fitting}

\item{...}{other parameters to be passed to plot method}
}
\description{
Plotting method for medde objects
}
