% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.B2mix.R
\name{predict.B2mix}
\alias{predict.B2mix}
\title{Predict Method for Bmix}
\usage{
\method{predict}{B2mix}(object, newdata, Loss = 2, newk, ...)
}
\arguments{
\item{object}{fitted object of class "B2mix"}

\item{newdata}{Values at which prediction is desired an n by 2 matrix}

\item{Loss}{Loss function used to generate prediction:  Currently supported values:
2 to get mean predictions, 1 to get median predictions, 0 to get modal predictions
or any tau in (0,1) to get tau-th quantile predictions.}

\item{newk}{k values (number of trials) for the predictions an n by 2 matrix}

\item{...}{optional arguments to predict}
}
\value{
A vector of predictions
}
\description{
Predict Method for Binomial Mixtures
}
\details{
The predict method for B2mix objects will compute posterior means.
}
\author{
Jiaying Gu and Roger Koenker
}
\keyword{nonparametric}
