% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.Bmix.R
\name{predict.Bmix}
\alias{predict.Bmix}
\title{Predict Method for Bmix}
\usage{
\method{predict}{Bmix}(object, newdata, Loss = 2, newk, ...)
}
\arguments{
\item{object}{fitted object of class "Bmix"}

\item{newdata}{Values at which prediction is desired}

\item{Loss}{Loss function used to generate prediction:  Currently supported values:
2 to get mean predictions, 1 to get median predictions, 0 to get modal predictions
or any tau in (0,1) to get tau-th quantile predictions.}

\item{newk}{k values (number of trials) for the predictions}

\item{...}{optional arguments to predict}
}
\value{
A vector of predictions
}
\description{
Predict Method for Binomial Mixtures
}
\details{
The predict method for \code{Bmix} objects will compute means, quantiles or
modes of the posterior according to the \code{Loss} argument.  Typically,
\code{newdata} would be passed to \code{predict}
}
\author{
Jiaying Gu
}
\keyword{nonparametric}
