% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.GLmix}
\alias{predict.GLmix}
\title{Predict Method for GLmix}
\usage{
\method{predict}{GLmix}(object, newdata, Loss = 2, ...)
}
\arguments{
\item{object}{fitted object of class "GLmix"}

\item{newdata}{Values at which prediction is desired}

\item{Loss}{Loss function used to generate prediction}

\item{...}{optional arguments to predict}
}
\value{
A vector of predictions
}
\description{
Predict Method for Gaussian Location Mixtures
}
\details{
The predict method for \code{GLmix} objects will compute means, medians or
modes of the posterior according to the \code{Loss} argument.  Typically,
\code{newdata} would be passed to \code{predict}
}
\author{
Roger Koenker
}
\keyword{nonparametric}

