% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.GLVmix.R
\name{predict.GLVmix}
\alias{predict.GLVmix}
\title{Predict Method for GLVmix}
\usage{
\method{predict}{GLVmix}(object, newdata, Loss = 2, ...)
}
\arguments{
\item{object}{Fitted object of class "GLVmix"}

\item{newdata}{data.frame with components(t,s,m) at which prediction is desired}

\item{Loss}{Loss function used to generate prediction:  Currently supported values: 
2 to get mean predictions, 1 to get median predictions, 0 to get modal predictions
or any tau in (0,1) to get tau-th quantile predictions.}

\item{...}{optional arguments to predict}
}
\value{
A vector of predictions
}
\description{
Predict Method for Gaussian Location-scale Mixtures
}
\details{
The predict method for \code{GLmix} objects will compute means, quantiles or
modes of the posterior according to the \code{Loss} argument.  Typically,
\code{newdata} would be passed to \code{predict}.  Note that these predictions
are for the location parameter only.
}
\author{
Roger Koenker
}
\keyword{nonparametric}
