% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLmix.R
\name{predict.HLmix}
\alias{predict.HLmix}
\title{Predict Method for HLmix}
\usage{
\method{predict}{HLmix}(object, newdata, Loss = 2, newsigma = NULL, ...)
}
\arguments{
\item{object}{fitted object of class "HLmix"}

\item{newdata}{Values at which prediction is desired}

\item{Loss}{Loss function used to generate prediction:  Currently supported values: 
2 to get mean predictions, 1 to get median predictions, 0 to get modal predictions
or any tau in (0,1) to get tau-th quantile predictions.}

\item{newsigma}{sigma values for the predictions}

\item{...}{optional arguments to predict}
}
\value{
A vector of predictions
}
\description{
Predict Method for Huber Location Mixtures
}
\details{
The predict method for \code{HLmix} objects computes means, quantiles or
modes of the posterior according to the \code{Loss} argument.  Typically,
\code{newdata} would be passed to \code{predict}.  Note that if newdata
is simply equal to the original observations (denoising case) then the
}
\author{
Roger Koenker
}
