% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.WGLVmix.R
\name{predict.WGLVmix}
\alias{predict.WGLVmix}
\title{Predict Method for WGLVmix}
\usage{
\method{predict}{WGLVmix}(object, newdata, Loss = 2, ...)
}
\arguments{
\item{object}{Fitted object of class "GLVmix"}

\item{newdata}{data.frame with components(y,id,w) at which prediction is desired
this data structure must be compatible with that of \code{WGLVmix}, if newdata$w
is NULL then w is replaced by a vector of ones of length(y)}

\item{Loss}{Loss function used to generate prediction:  Currently supported values: 
2 to get mean predictions, 1 to get median predictions, 0 to get modal predictions
or any tau in (0,1) to get tau-th quantile predictions.}

\item{...}{optional arguments to predict}
}
\value{
A vector of predictions
}
\description{
Predict Method for Gaussian Location-scale Mixtures (Longitudinal Version)
}
\details{
The predict method for \code{WGLmix} objects will compute means, quantiles or
modes of the posterior according to the \code{Loss} argument.  Typically,
\code{newdata} would be passed to \code{predict}.  Note that these predictions
are for the location parameter only.
}
\author{
Roger Koenker
}
\keyword{nonparametric}
