% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychtest.R
\name{psychtest}
\alias{psychtest}
\title{Lord and Cressie Binomial Psychological Test Data}
\format{
A data frame with 21 observations on 2 variables.
\describe{
    \item{\code{x}}{number of correct answers (out of 20 questions)}
    \item{\code{k}}{frequency out of 12990 exams.}
}
}
\source{
Lord, F.M. and N. Cressie  (1975).  An Empirical Bayes Procedure for 
Finding an Interval Estimate, Sankhya, 37, 1-9.
}
\usage{
psychtest
}
\description{
Frequencies of test scores (number of correct answers) 
on a psychological test with 20 questions.
}
\keyword{datasets}
