% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{qKW}
\alias{qKW}
\title{Quantiles of KW fit}
\usage{
qKW(g, q)
}
\arguments{
\item{g}{KW fitted object}

\item{q}{vector of quantiles to be computed}
}
\description{
Quantiles of KW fit
}
\author{
R. Koenker
}
\keyword{utility}
