% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{qKW2}
\alias{qKW2}
\title{Quantiles of bivariate KW fit}
\usage{
qKW2(g, q)
}
\arguments{
\item{g}{KW fitted object}

\item{q}{vector of quantiles to be computed}
}
\description{
Quantiles of bivariate KW fit
}
\author{
R. Koenker
}
\keyword{utility}
