% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medde.R
\name{qmedde}
\alias{qmedde}
\title{Quantile function for medde estimate}
\usage{
qmedde(p, medde)
}
\arguments{
\item{p}{vector of probabilities at which to evaluate the quantiles}

\item{medde}{fitted object from medde}
}
\description{
Slightly modified version  borrowed from the package logcondens 
Todo:  extend this to cases with \eqn{\alpha != 1}.
}
\keyword{nonparametric}
