% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rKW}
\alias{rKW}
\title{Random sample from KW object}
\usage{
rKW(n, g)
}
\arguments{
\item{n}{sample size}

\item{g}{KW object}
}
\description{
Random sample from KW object
}
\author{
R. Koenker
}
\keyword{utility}
