% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medde.R
\name{rmedde}
\alias{rmedde}
\title{Random number generation from a medde estimate}
\usage{
rmedde(n, medde, smooth = TRUE)
}
\arguments{
\item{n}{number of observations desired in calls to rmedde}

\item{medde}{fitted medde object for calls in qmedde and rmedde}

\item{smooth}{option to draw random meddes from the smoothed density}
}
\description{
Random number generation from a medde estimate
}
\keyword{nonparametric}

