% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ThreshFDR}
\alias{ThreshFDR}
\title{Thresholding for False Discovery Rate}
\usage{
ThreshFDR(lambda, stat, v)
}
\arguments{
\item{lambda}{is the proposed threshold}

\item{stat}{is the statistic used for ranking}

\item{v}{is the local false discovery statistic}
}
\description{
This function approximates FDR for various values of lambda
and is usually employed in conjunction with \code{Finv} to
find an appropriate cutoff value lambda.
}
\keyword{utility}
