\name{traprule}
\alias{traprule}
\title{ Integration by Trapezoidal Rule }
\description{ Integration by Trapezoidal Rule }
\usage{ traprule(x, y) }
\arguments{
  \item{x}{ points of evaluation }
  \item{y}{ function values }
}
\details{ Crude Riemann sum approximation.  }
\value{ A real number.  }
\author{ R. Koenker }

\keyword{ utility}
