% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/medde.R
\name{plot.medde}
\alias{plot.medde}
\title{Plotting method for medde objects}
\usage{
\method{plot}{medde}(x, xlab = "x", ylab = "f(x)", ...)
}
\arguments{
\item{x}{object obtained from medde fitting}

\item{xlab}{label for horizontal axis}

\item{ylab}{label for vertical axis}

\item{...}{other parameters to be passed to plot method}
}
\description{
Plotting method for medde objects
}

