% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{traprule}
\alias{traprule}
\title{Integration by Trapezoidal Rule}
\usage{
traprule(x, y)
}
\arguments{
\item{x}{points of evaluation}

\item{y}{function values}
}
\value{
A real number.
}
\description{
Integration by Trapezoidal Rule
}
\details{
Crude Riemann sum approximation.
}
\author{
R. Koenker
}
\keyword{utility}

